requirements.txt
aiocontextvars==0.2.2
allure-pytest==2.8.6
allure-python-commons==2.8.6
APScheduler==3.6.3
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
colorlog==4.1.0
contextvars==2.4
cryptography==2.9.2
dataclasses==0.7
filetype==1.0.7
fixedint==0.1.5
gevent==1.4.0
greenlet==0.4.15
har2case==0.3.1
httprunner==2.5.7
idna==2.7
immutables==0.14
importlib-metadata==1.3.0
itsdangerous==1.1.0
jsonpath==0.82
jsonschema==3.2.0
jwt==1.0.0
loguru==0.4.1
MarkupSafe==1.1.1
more-itertools==8.0.2
paramiko==2.7.1
pip==20.1.1
pluggy==0.13.1
protobuf==3.11.3
psutil==5.6.2
py==1.8.0
pycparser==2.20
pydantic==1.5.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyrsistent==0.16.0
pytest==4.5.0
pytest-html==1.19.0
pytest-metadata==1.8.0
PyTestReport==0.1.9.3
python-dateutil==2.8.1
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2019.1
PyYAML==5.2
requests==2.19.1
requests-toolbelt==0.8.0
setuptools==40.6.2
SQLAlchemy==1.3.15
urllib3==1.23
